# specs/002-backend-api/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: Backend API for the Todo Full-Stack Web Application - Phase II
  version: 1.0.0

servers:
  - url: http://localhost:7860
    description: Local development server

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      description: Public endpoint to verify backend availability
      security: []
      responses:
        '200':
          description: Backend is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  /api/{user_id}/tasks:
    get:
      summary: List all tasks for a user
      description: Returns all tasks owned by the authenticated user. URL user_id must match JWT user_id.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID (must match JWT sub claim)
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieves a specific task by ID, scoped to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved the task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update a task
      description: Updates an existing task owned by the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete a task
      description: Deletes a task owned by the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggles the completion status of a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          description: The title of the task
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          description: Optional description
          example: "Milk, bread, eggs"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        completed:
          type: boolean

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
        user_id:
          type: string
          description: Owner user ID
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

  responses:
    BadRequestError:
      description: Invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Validation error"
              detail:
                type: string

    UnauthorizedError:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Authentication required"

    ForbiddenError:
      description: User ID mismatch - access denied
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Access forbidden"

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Task not found"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth